March 29th working

+++ projects/web-frontend/src/schemas/odlc.ts
@@ -6,7 +6,7 @@ export const OdlcImageSchema = z.object({
     // So the PNG isn't an image but instead each pixel is a 16-bit unsigned integer representing depth in millimeters. We can decode this to get depth values.
     color_detection: z.tuple([z.number().int(), z.number().int(), z.number().int()]),
     bounding_box: z.array(z.tuple([z.number(), z.number()])),
-    confidence_level: z.number().int(),
+    confidence_level: z.number(),
 });
 
 export const OdlcSessionPayloadSchema = z.object({
